tg_parser/requirements.txt
python-telegram==0.18.0
setuptools==68.2.2
telegram-text==0.1.2
